Fix upload/download treatments
