+++ xDrip/Model/Entries/AbstractEntry.swift
@@ -18,6 +18,10 @@ class AbstractEntry: Object {
     @objc private(set) dynamic var date: Date?
     @objc private(set) dynamic var externalID: String?
     @objc private dynamic var rawCloudUploadStatus: Int = CloudUploadStatus.notApplicable.rawValue
+    
+    override class func primaryKey() -> String? {
+        return "externalID"
+    }
 
     var cloudUploadStatus: CloudUploadStatus {
         get {
